pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web2vault"
version = "0.1.0"
description = "Scrape web pages and generate Obsidian-compatible study notes using LLMs"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "firecrawl-py>=4.0",
    "anthropic>=0.70",
    "openai>=1.50",
    "python-dotenv>=1.0",
]

[project.scripts]
web2vault = "web2vault.cli:main"
